refactor: [SNOW-1942858] move mock_connect to fixtures.py
